+++ src/libdrakvuf/linux-processes.c
@@ -222,7 +222,7 @@ bool linux_get_process_pid(drakvuf_t drakvuf, addr_t process_base, vmi_pid_t* pi
     return false;
 }
 
-status_t linux_get_process_tid(drakvuf_t drakvuf, addr_t process_base, vmi_pid_t* tid )
+bool linux_get_process_tid(drakvuf_t drakvuf, addr_t process_base, vmi_pid_t* tid )
 {
     /*
      * On Linux PID is actually the thread ID....... ... ...
@@ -234,7 +234,10 @@ status_t linux_get_process_tid(drakvuf_t drakvuf, addr_t process_base, vmi_pid_t
         .addr = process_base + drakvuf->offsets[TASK_STRUCT_PID]
     };
 
-    return vmi_read_32(drakvuf->vmi, &ctx, (uint32_t*)tid);
+    if ( VMI_SUCCESS == vmi_read_32(drakvuf->vmi, &ctx, (uint32_t*)tid))
+        return true;
+
+    return false;
 }
 char* linux_get_current_process_name(drakvuf_t drakvuf, drakvuf_trap_info_t* info, bool fullpath)
 {
@@ -363,15 +366,21 @@ bool linux_get_process_data( drakvuf_t drakvuf, addr_t base_addr, proc_data_priv
     if ( !base_addr )
         return false;
 
-            if ( proc_data->name )
-            {
-                proc_data->userid = linux_get_process_userid( drakvuf, base_addr );
-                linux_get_process_ppid( drakvuf, base_addr, &proc_data->ppid );
-                linux_get_process_tid(drakvuf, base_addr, &proc_data->tid);
-                return true;
-            }
-        }
-    }
+    if ( !linux_get_process_pid(drakvuf, base_addr, &proc_data->pid) )
+        return false;
+
+    proc_data->name = linux_get_process_name(drakvuf, base_addr, true);
+
+    if ( !proc_data->name )
+        return false;
+
+    proc_data->userid = linux_get_process_userid(drakvuf, base_addr);
+
+    if ( !linux_get_process_ppid(drakvuf, base_addr, &proc_data->ppid) )
+        PRINT_DEBUG("Failed to gather parent process' PID for %s:%u\n", proc_data->name, proc_data->pid);
+
+    if (!linux_get_process_tid(drakvuf, base_addr, &proc_data->tid))
+        PRINT_DEBUG("Failed to gather current process tid");
 
     return false;
 }
