+++ src/libdrakvuf/linux.c
@@ -148,7 +148,7 @@ bool set_os_linux(drakvuf_t drakvuf)
     drakvuf->osi.get_process_tid = linux_get_process_tid;
     drakvuf->osi.get_process_ppid = linux_get_process_ppid;
     drakvuf->osi.get_process_data = linux_get_process_data;
-    drakvuf->osi.export_libc_address = linux_get_address_of_libc;
+    drakvuf->osi.export_linux_sym_to_va = process_sym2va;
 
     return 1;
 }
