+++ src/libdrakvuf/linux.h
@@ -116,7 +116,7 @@ char* linux_get_process_name(drakvuf_t drakvuf, addr_t process_base, bool fullpa
 
 bool linux_get_process_pid(drakvuf_t drakvuf, addr_t process_base, vmi_pid_t* pid);
 
-status_t linux_get_process_tid(drakvuf_t drakvuf, addr_t process_base, vmi_pid_t* tid);
+bool linux_get_process_tid(drakvuf_t drakvuf, addr_t process_base, vmi_pid_t* tid);
 
 char* linux_get_current_process_name(drakvuf_t drakvuf, drakvuf_trap_info_t* info, bool fullpath);
