+++ src/libdrakvuf/os.h
@@ -115,8 +115,11 @@ typedef struct os_interface
     addr_t (*get_current_process)
     (drakvuf_t drakvuf, drakvuf_trap_info_t* info);
 
-    addr_t (*export_libc_address)
-    (drakvuf_t drakvuf, drakvuf_trap_info_t* info, vmi_pid_t pid, const char* libc);
+    bool (*get_last_error)
+    (drakvuf_t drakvuf, drakvuf_trap_info_t* info, uint32_t* err, const char** err_str);
+
+    addr_t (*export_linux_sym_to_va)
+    (drakvuf_t drakvuf, drakvuf_trap_info_t* info, vmi_pid_t pid, const char* lib);
 
     status_t (*get_last_error)
     (drakvuf_t drakvuf, drakvuf_trap_info_t* info, uint32_t* err, const char** err_str);
@@ -136,7 +139,7 @@ typedef struct os_interface
     bool (*get_process_pid)
     (drakvuf_t drakvuf, addr_t process_base, vmi_pid_t* pid);
 
-    status_t (*get_process_tid)
+    bool (*get_process_tid)
     (drakvuf_t drakvuf, addr_t process_base, vmi_pid_t* tid);
 
     int64_t (*get_current_process_userid)
