+++ src/libdrakvuf/private.h
@@ -275,12 +275,12 @@ struct remapped_gfn
 
 typedef struct process_data_priv
 {
-    char* name;   /* Process name */
-    vmi_pid_t pid ;     /* Process pid | tid in linux*/
+    char* name;         /* Process name */
+    vmi_pid_t pid ;     /* Process pid | tgid in linux*/
     vmi_pid_t ppid ;    /* Process parent pid */
     addr_t base_addr ;  /* Process base address */
     int64_t userid ;    /* Process SessionID/UID */
-    vmi_pid_t tid; /* Thread id for Linux*/
+    vmi_pid_t tid;      /* Thread id for Linux*/
 } proc_data_priv_t ;
 
 struct memcb_pass
