+++ src/libinjector/linux_private.h
@@ -125,68 +125,4 @@ extern bool verbose;
 #define vmi_free_unicode_str g_free
 #endif
 
-enum offset
-{
-    KTHREAD_TRAPFRAME,
-    KTRAP_FRAME_RIP,
-
-    OFFSET_MAX
-};
-
-static const char* offset_names[OFFSET_MAX][2] =
-{
-    [KTHREAD_TRAPFRAME] = {"_KTHREAD", "TrapFrame" },
-    [KTRAP_FRAME_RIP] = {"_KTRAP_FRAME", "Rip" },
-};
-
-enum linux_offset
-{
-
-
-    LINUX_OFFSET_MAX
-};
-
-static const char* linux_offset_names[LINUX_OFFSET_MAX][2] =
-{
-
-
-
-};
-
-struct pt_regs
-{
-    /*
-     * C ABI says these regs are callee-preserved. They aren't saved on kernel entry
-     * unless syscall needs a complete, fully filled "struct pt_regs".
-     */
-    unsigned long r15;
-    unsigned long r14;
-    unsigned long r13;
-    unsigned long r12;
-    unsigned long bp;
-    unsigned long bx;
-    /* These regs are callee-clobbered. Always saved on kernel entry. */
-    unsigned long r11;
-    unsigned long r10;
-    unsigned long r9;
-    unsigned long r8;
-    unsigned long ax;
-    unsigned long cx;
-    unsigned long dx;
-    unsigned long si;
-    unsigned long di;
-    /*
-     * On syscall entry, this is syscall#. On CPU exception, this is error code.
-     * On hw interrupt, it's IRQ number:
-     */
-    unsigned long orig_ax;
-    /* Return frame for iretq */
-    unsigned long ip;
-    unsigned long cs;
-    unsigned long flags;
-    unsigned long sp;
-    unsigned long ss;
-    /* top of stack page */
-};
-
 #endif
